+++ chatbotkit/__init__.py
@@ -1,12 +1,19 @@
 from ._client import ChatBotKit
-from ._transport import APIError, Client, ClientOptions, Response
+from ._transport import (
+    APIError,
+    AuthorizationRequiredError,
+    Client,
+    ClientOptions,
+    Response,
+)
 
 __all__ = [
     "APIError",
+    "AuthorizationRequiredError",
     "ChatBotKit",
     "Client",
     "ClientOptions",
     "Response",
 ]
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"